+++ nix/modules/project/settings/default.nix
@@ -50,7 +50,7 @@ in
       let
         applySettingsFor = name: mod:
           let
-            cfg = (lib.evalModules {
+            cfg' = (lib.evalModules {
               modules = [
                 # Settings spec
                 ./all.nix
@@ -71,13 +71,19 @@ in
                 config = cfg;
               });
             }).config;
+            cfg = traceSettings name cfg';
+            # HACK: buildFromSdist must apply *last*
+            # cf. https://github.com/srid/haskell-flake/pull/252
+            # In future, we can refactor this as part of https://github.com/srid/haskell-flake/issues/285
+            impl = lib.attrsets.removeAttrs cfg.impl [ "buildFromSdist" ];
+            fns = lib.attrValues impl ++ [ cfg.impl.buildFromSdist ];
           in
           lib.pipe super.${name} (
             # TODO: Do we care about the *order* of overrides?
             # Might be relevant for the 'custom' option.
             lib.concatMap
               (impl: impl)
-              (lib.attrValues (traceSettings name cfg).impl)
+              fns
           );
       in
       lib.mapAttrs applySettingsFor project.config.settings;